fix: 棚卸し完了したときに死ぬ問題を解消
useMappedStateはstateが更新された瞬間に全ての場所で即時評価される。
そのため、eventがDoneになった時点でInventoryEventDoing内で即時useMappedStateが評価される
このとき、Doingとみなしていた箇所で実行事例街になる
