+++ packages/front/src/bibliotheca/features/inventory/inventoryEvent/components/InventoryDoing.tsx
@@ -1,55 +1,27 @@
 import { Link } from 'bibliotheca/components/Link';
 import { StyledDataTable } from 'bibliotheca/components/StyledDataTable';
-import { findUncheckedOnlyList } from 'bibliotheca/services/inventory/query';
-import { InventoryEventDoing, Book } from 'bibliotheca/types';
+import { Book, InventoryStatus } from 'bibliotheca/types';
 import { Button, RadioButton, Text } from 'grommet';
 import React from 'react';
-import { useActions, useMappedState } from 'typeless';
-import { getInventoryBookModuleState } from '../../inventoryBookModule/interface';
-import { getInventoryEventState, InventoryEventActions } from '../interface';
+import { useActions } from 'typeless';
+import { InventoryEventActions, ViewType } from '../interface';
 
-export const InventoryDoing = () => {
+type BookForTable = Book & { status: InventoryStatus; key: number };
+type InventoryEventDoingProps = {
+  canChangeMissingAll: boolean;
+  books: BookForTable[];
+  viewType: ViewType;
+};
+
+export const InventoryDoing = ({
+  canChangeMissingAll,
+  viewType,
+  books,
+}: InventoryEventDoingProps) => {
   const { changeView, toMissingAll, submitInventory, toCheckStatus } = useActions(
     InventoryEventActions,
   );
-  const { canChangeMissingAll, books, viewType } = useMappedState(
-    [getInventoryBookModuleState, getInventoryEventState],
-    ({ booksInList, event }, { viewType }) => {
-      const uncheckedBooks = findUncheckedOnlyList(
-        (event as InventoryEventDoing).inventoryBooks,
-        booksInList,
-      );
-      const canChangeMissingAll = uncheckedBooks.length === 0;
-      const books = ((e: InventoryEventDoing) => {
-        switch (viewType) {
-          case 'checkedOnly':
-            return e.inventoryBooks.map(({ status, bookId }) => ({
-              status,
-              ...booksInList.find(b => b.id === bookId)!,
-            }));
-          case 'all':
-            return booksInList.map(b => {
-              const inventoryBook = e.inventoryBooks.find(ib => ib.bookId === b.id);
-              if (inventoryBook) {
-                return { status: inventoryBook.status, ...b };
-              } else {
-                return { status: 'unchecked', ...b };
-              }
-            });
-          case 'uncheckedOnly': {
-            return uncheckedBooks;
-          }
-          default:
-            throw new Error('unknown mode');
-        }
-      })(event as InventoryEventDoing);
-      return {
-        viewType,
-        books: books.map((b, i) => ({ ...b, key: i })),
-        canChangeMissingAll,
-      };
-    },
-  );
+
   return (
     <>
       <Link href={`/inventory-event/register-book`}>
